Bugs fixes

+++ CareConnect.WPF/ViewModels/UserControls/UsersAdminViewModel.cs
@@ -61,6 +61,16 @@ namespace CareConnect.WPF.ViewModels.UserControls
 
         private void DeleteUser(object? paramater)
         {
+            var verifyString = _usersAdminService.VerifyLinkingObjects(SelectedUser!.Id);
+
+            if (string.Equals(verifyString, "bookings", StringComparison.OrdinalIgnoreCase))
+                if (MessageBox.Show("This client user has active bookings. Are you sure you want to proceed?", "Warning", MessageBoxButton.YesNo) == MessageBoxResult.No)
+                    return;
+
+            if (string.Equals(verifyString, "services", StringComparison.OrdinalIgnoreCase))
+                if (MessageBox.Show("This staff user has active services. Are you sure you want to proceed?", "Warning", MessageBoxButton.YesNo) == MessageBoxResult.No)
+                    return;
+
             _usersAdminService.DeleteUser(SelectedUser!.Id, _activeUserService.ActiveUser!.Id);
             UsersList = new ObservableCollection<User>(_usersAdminService.GetUsers(SelectedTypeUsers));
